memory usage type

+++ header/actralloc.h
@@ -22,6 +22,6 @@ void *actr_malloc(int size)
 /// @brief returns a simple usage memory report
 /// @return allocated string must be freed
 extern char *actr_memory_report();
-extern double actr_memory_usage();
+extern long long actr_memory_usage();
 
 #endif
